# Improve ucaps.h `_find_upper()` and `_find_lower()` performance