Updated `_find_upper()` and `_find_lower()` to use a hashmap-like lookup instead of binary search.  Reduces time of functions like `String::findn()`, `String::to_upper()`, and `String::to_lower()` by around 50-80% depending on strings used.

Compared using gdscript below:

```
extends Node2D

var text := "Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat. Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur. Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum."

func _ready() -> void:
	test_findn()
	test_to_lower()
	test_to_upper()

func test_findn():
	var ms_start = Time.get_ticks_msec()
	var to_find := "abc"
	for i in 1000000:
		text.findn(to_find)
	var ms_end = Time.get_ticks_msec()
	print("findn: " + str(ms_end - ms_start) + "ms")

func test_to_lower():
	var ms_start = Time.get_ticks_msec()
	for i in 1000000:
		text.to_lower()
	var ms_end = Time.get_ticks_msec()
	print("to_lower: " + str(ms_end - ms_start) + "ms")

func test_to_upper():
	var ms_start = Time.get_ticks_msec()
	for i in 1000000:
		text.to_upper()
	var ms_end = Time.get_ticks_msec()
	print("to_upper: " + str(ms_end - ms_start) + "ms")

```

Old:
```
findn: 13111ms
to_lower: 6793ms
to_upper: 5231ms
```

New:
```
findn: 2428ms
to_lower: 1435ms
to_upper: 2241ms
```

<!--
Please target the `master` branch in priority.

Relevant fixes are cherry-picked for stable branches as needed by maintainers.

To speed up the contribution process and avoid CI errors, please set up pre-commit hooks locally:
https://docs.godotengine.org/en/latest/contributing/development/code_style_guidelines.html
-->