@AThousandShips 2024-06-19
See also:
* https://github.com/godotengine/godot/pull/90036