@Ivorforce 2025-01-05
Just to state it here too: I have opened https://github.com/godotengine/godot/pull/99971, which can yield better performance benefits for latin texts.
That's not to say there's no value in this PR! #99971 _only_ optimizes latin (ascii+) texts. Texts with non-latin contents, like chinese, would benefit from this PR. But we should be aware of this caveat when considering merging this :)